--- src/pages/dashboard/daily.page.jsx
@@ -1,5 +1,0 @@
-import { DailyContainer } from "@features/dashboard/daily";
-
-export default function DailyPage() {
-  return <DailyContainer />;
-}
