+++ src/pages/dashboard/dashboard.routes.jsx
@@ -1,11 +1,9 @@
-import { Routes, Route } from "react-router-dom";
+import { Routes, Route, Navigate } from "react-router-dom";
 import { ProtectedRoute } from "@features/users/auth/protectedRoute";
 import { UserPreferenceProvider } from "@context/userPreference.context";
 import DashboardPage from "./dashboard.page";
 import LockerPage from "./locker.page";
 import DatalabPage from "./datalab.page";
-import DailyPage from "./daily.page";
-import MonthlyPage from "./monthly.page";
 import ScheduleRoutes from "./schedule/schedule.routes";
 import PlayingRoutes from "./playing/playing.routes";
 
@@ -18,8 +16,14 @@ export default function DashboardRoutes() {
             <Route path="/" element={<DashboardPage />} />
             <Route path="/locker" element={<LockerPage />} />
             <Route path="/datalab" element={<DatalabPage />} />
-            <Route path="/daily" element={<DailyPage />} />
-            <Route path="/monthly" element={<MonthlyPage />} />
+            <Route
+              path="/daily"
+              element={<Navigate to="/dashboard/schedule/daily" replace />}
+            />
+            <Route
+              path="/monthly"
+              element={<Navigate to="/dashboard/schedule/monthly" replace />}
+            />
             <Route path="/schedule/*" element={<ScheduleRoutes />} />
             {/* <Route path="/playing/*" element={<PlayingRoutes />} /> */}
           </Routes>
