+++ src/pages/dashboard/schedule/monthlypage.page.jsx
@@ -1,6 +1,7 @@
 import { useState, useEffect } from "react";
 import { getData } from "@utils/api/data";
 import MonthlyMatches from "@features/dashboard/monthly/monthly-matches";
+import MonthlySummary from "@features/dashboard/monthly/monthly-summary";
 import { Button } from "@ui/button";
 import { ChevronLeft, ChevronRight } from "lucide-react";
 
@@ -14,30 +15,77 @@ export default function MonthlyPage() {
   const fetchMonthlyData = async () => {
     setLoading(true);
     try {
-      const formattedDate = `${selectedDate.getFullYear()}-${String(
-        selectedDate.getMonth() + 1
-      ).padStart(2, "0")}`;
+      // S'assurer que le mois est sur 2 chiffres
+      const month = String(selectedDate.getMonth() + 1).padStart(2, "0");
+      const year = selectedDate.getFullYear();
+      const formattedDate = `${year}-${month}`;
 
-      console.log("🚀 Fetching data for:", formattedDate);
-      const response = await getData(
-        `v1/matches/monthly_metrics/${formattedDate}`
+      console.log("📅 Date formatée:", formattedDate);
+      console.log("🚀 Fetching monthly summary for:", formattedDate);
+      const monthlyResponse = await getData(
+        `v1/summaries/monthly/${formattedDate}`
       );
-      console.log("📦 API Response:", response);
+      console.log("📦 Monthly API Response:", monthlyResponse);
 
-      if (response?.daily_metrics) {
-        console.log("✅ Setting daily metrics:", response.daily_metrics);
-        setDailyMetrics(response.daily_metrics);
+      console.log("🚀 Fetching monthly matches for:", formattedDate);
+      const matchesResponse = await getData(
+        `v1/matches/monthly/${formattedDate}`
+      );
+      console.log("📦 Matches API Response:", matchesResponse);
+
+      if (monthlyResponse) {
+        // Transformer la réponse pour correspondre à la structure attendue
+        const monthlyTotals = {
+          total_matches: monthlyResponse.total_matches,
+          total_energy: monthlyResponse.total_energy,
+          total_bft: monthlyResponse.total_bft.amount,
+          total_flex: monthlyResponse.total_flex.amount,
+          profit: monthlyResponse.profit,
+          total_wins: monthlyResponse.results.win,
+          total_losses: monthlyResponse.results.loss,
+          total_draws: monthlyResponse.results.draw,
+          win_rate: monthlyResponse.results.win
+            ? (
+                (monthlyResponse.results.win / monthlyResponse.total_matches) *
+                100
+              ).toFixed(1)
+            : "0.0",
+        };
+
+        console.log("✅ Setting monthly totals:", monthlyTotals);
+        setMonthlyTotals(monthlyTotals);
       } else {
-        console.warn("⚠️ No daily metrics in response");
-        setDailyMetrics({});
+        console.warn("⚠️ No monthly response from API");
+        setMonthlyTotals({});
       }
 
-      if (response?.monthly_totals) {
-        console.log("✅ Setting monthly totals:", response.monthly_totals);
-        setMonthlyTotals(response.monthly_totals);
+      if (matchesResponse?.matches) {
+        // Les données sont déjà groupées par jour dans la réponse
+        const dailyMetrics = {};
+
+        // Transformer chaque jour pour correspondre à notre format
+        Object.entries(matchesResponse.matches).forEach(([date, dayData]) => {
+          dailyMetrics[date] = {
+            total_matches: dayData.total_matches,
+            total_energy: dayData.total_energy,
+            total_bft: dayData.total_bft.amount,
+            total_flex: dayData.total_flex.amount,
+            wins: dayData.results.win,
+            losses: dayData.results.loss,
+            draws: dayData.results.draw,
+            total_energy_cost: dayData.total_energy_cost,
+            total_bft_value: dayData.total_bft.value,
+            total_flex_value: dayData.total_flex.value,
+            total_profit: dayData.profit,
+            win_rate: dayData.win_rate,
+          };
+        });
+
+        console.log("✅ Setting daily metrics:", dailyMetrics);
+        setDailyMetrics(dailyMetrics);
       } else {
-        console.warn("⚠️ No monthly totals in response");
-        setMonthlyTotals({});
+        console.warn("⚠️ No matches in response");
+        setDailyMetrics({});
       }
     } catch (error) {
       console.error("❌ Error fetching monthly data:", error);
@@ -83,6 +131,10 @@ export default function MonthlyPage() {
             <ChevronRight className="h-4 w-4" />
           </Button>
         </div>
+      </div>
+
+      <div className="px-5 mb-6">
+        <MonthlySummary date={selectedDate} metrics={monthlyTotals} />
       </div>
 
       <MonthlyMatches
